+++ packages/svelte/src/lib/components/checkbox/use-checkbox-group-context.ts
@@ -3,7 +3,7 @@ import type { UseCheckboxGroupReturn } from './use-checkbox-group.svelte'
 
 export interface UseCheckboxGroupContext extends UseCheckboxGroupReturn {}
 
-export const [CheckboxGroupProvider, useCheckboxGroupContext] = createContext<UseCheckboxGroupContext>({
+export const [CheckboxGroupContextProvider, useCheckboxGroupContext] = createContext<UseCheckboxGroupContext>({
   name: 'CheckboxGroupContext',
   strict: false,
 })